Sync LLC Java snippets
