+++ stream-chat-android-docs/src/main/java/io/getstream/chat/docs/java/client/cms/UserPresence.java
@@ -1,4 +1,4 @@
-package io.getstream.chat.docs.java;
+package io.getstream.chat.docs.java.client.cms;
 
 import java.util.Arrays;
 import java.util.List;
